requirements.txt
django==4.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
django-filter==22.1
mysqlclient==2.1.1
Pillow==9.2.0
autopep8==1.7.0
psycopg2-binary==2.9.5
postgres==4.0